Extract of main function in compiler
