+++ src/qry/qryLauncher.cpp
